+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.51.0](https://github.com/nexxtway/rainbow-modules/compare/v0.50.0...v0.51.0) (2022-06-28)
+
+### Features
+
+-   add waiting time icon ([#561](https://github.com/nexxtway/rainbow-modules/issues/561)) ([0021dcc](https://github.com/nexxtway/rainbow-modules/commit/0021dcc17c15fc1066e434a6325a0a48d42ccbae))
+
 # [0.50.0](https://github.com/nexxtway/rainbow-modules/compare/v0.49.0...v0.50.0) (2022-04-22)
 
 ### Bug Fixes